Switched authentication to YesodAuthEmail
Part of this commit fixes some problems with the stock Yesod.Auth.Email
behavior.

There was no "forgot password" link on the login form, fixed that. Also,
wording on some of the buttons/labels, changed to reflect what we need
for this project.

The Forgot Password handler was allowing both an identifier and email
address. This project will be using email addresses as the sole
identifier for users, so that was removed.

In addition: Reorganized Foundation LANGUAGE and import order. I like
these things to be arranged alphabetically.