Back to using Yesod.Auth.Email for authentication
We think this will service the needs of this web app best and brings a
lot of good default behavior (like forgot password reset and email
verification) to the table.

Part of this change was also upgrading to Stackage lts-11.8 which brings
us up to Yesod 1.6.x. This is a big change that was eventually going to
be necessary.