Fixed a bug lost form data values
I didn't understand that ALL of the fields in the data structures has to
pass through the form and come back after submit or be lost. As a
workaround, using hidden fields in my User forms. But this has
introduced layout problems.

Going to investigate using monadic forms as an alternative to this.