Added form and code to edit user accounts
Part of this code consolidates the Helper.UserEdit module into
Handler.UserEdit. It's not used anywhere else.

Also got rid of userFormEmpty because we always need to get the user record
being edited from the db anyway, may as well just use that.

Added isAdminB check to possibly disable editing of the Membership Number
field. We only allow this for administrators. In the future, users will be able
to edit their own records (but not that field).